getting 3.10 to run on windows, has temp debug info in pianosyn.lsp
git-svn-id: svn://svn.code.sf.net/p/nyquist/code/trunk/nyquist@209 7a394894-9432-4381-9535-169912d1abbc

+++ cmupv/src/internal.c
@@ -10,12 +10,12 @@
 
 float hann(double x)
 {
-    return 0.5 * (1.0 - cos(2 * M_PI * x));
+    return (float) (0.5 * (1.0 - cos(2 * M_PI * x)));
 }
 
 float hamm(double x)
 {
-    return 0.54 - 0.46 * cos(2 * M_PI * x);
+    return (float) (0.54 - 0.46 * cos(2 * M_PI * x));
 }
 
 void OneDimensionFFTshift(float vector[], int VectorLength)
